Add Featured Image Data Syndication
This will allow plugin to syndicate such fields as caption, alternative
text and description. It pulls post data and alt-text meat for the
syndicated image and then updates the remote post once image is
successfully uploaded and the attachment ID is created.
It would be easier to do it on the upload stage but xmlrpc media object
is not supporting caption and description fields at the moment.