Duplicate prevention also removes images on edit
This fixes the issue by preventing submitting metadata IDs that was not
updated during edit. This bug was introduced when issue #35 was fixed.

All custom field IDs being sent to remote site for deletion. However if
custom thumbnails were not edited and custom_thumb_id matches
_syn_thumb_thumb_id (syndicated reference) it's excluded form
processing and later is lost after deletion on remote. This will fix
the issue.