Settings page: add a maximum pull attempts option.
This option will be used by the plugin to disable sites that fail too many times in a row.

I've intentionally chosen to keep this option separate from the push_syndicate_settings option (which is an array of all plugin options). This is so it can be referenced more easily from other components without needing to get a hold of the WP_Push_Syndication_Server instance.

References issue #20.