46 - This exception logic I just implemented is wack
Added this late last night, it's a terrible idea, removing. Exceptions should really only be thrown when something 'exceptional' has occured, something outside of our control. The 'getting of a pull client' entails some assertion and that is expected, not exceptional. Let's KISS and handle them with a simple return false.