Fixes the "Pull Now" button in Settings Screen
Previously the “Pull Now” button only saved the settings.  We’re now
hooked into the settings validation (nasty, but a safe place to do it)
and will catch if settings have been saved, or if settings have been
saved and a pull has been requested.

If a pull has been requested, a single cron event for “now” is created
for each enabled site.