Declare Syndication globally in a more explicit way.
Some hosts load plugins differently, and this may impact the way that Syndication is available for reference in some circumstances (such as when running cron jobs), causing Syndication to not be available.

This change uses `$GLOBALS` instead of the `global` keyword to more reliably make Syndication available to plugins and themes.