pre_schedule_push_content update to include post_status check.
Currently Syndication is sending the posts on any save action regardless of if the Post is public. To minimize overhead we are restriting syndication further to only posts that have been syndicated already or are transition from/to a post_status that is approved.

Users can use the  filter to return an array of approved filters. Default array is simply [ 'publish']