Transient unique by POST->ID
Currently a single transient is used to determine if syndication is underway. If it is then syndication is cancelled for a post without retrying. To ensure we syndicate all content the transient is being updated to be unique by post so that the same post isn't syndicating when already in process.